requirements.txt
openai~=1.58.1
Flask~=3.1.0
google-generativeai~=0.8.3